tps/pyproject.toml
[project]
name = "tps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ansi2html==1.2.0",
    "celery>=5.2,<6",
    "django>=3.2,<4.0",
    "django-allauth>=0.51,<1",
    "django-bootstrap3==8.2.3",
    "django-clone",
    "django-debug-toolbar>=2.2.1",
    "django-extensions>=3.1.1",
    "django-import-export>=1.2.0",
    "django-multiselectfield==0.1.4",
    "django-redis<=4.7.0",
    "mock==2.0.0",
    "model-mommy==1.3.0",
    "pretty-bad-protocol>=3.1.1",
    "pygit2>=1.18.1",
    "python-magic>=0.4.27",
    "redis==2.10.6",
]

[tool.uv.sources]
django-clone = { git = "https://github.com/mohammadroghani/django-clone.git" }
